Fix 32bit warning, coding style.

+++ core/meta/src/TClass.cxx
@@ -175,10 +175,11 @@ Bool_t TClass::TDeclNameRegistry::HasDeclName(const char *name) const
 
 TClass::TDeclNameRegistry::~TDeclNameRegistry()
 {
-   if (fVerbLevel>1){
-      printf("TDeclNameRegistry Destructor. List of %lu names:\n",fClassNamesSet.size());
-      for(auto const & key:fClassNamesSet){
-         printf(" - %s\n",key.c_str());
+   if (fVerbLevel > 1) {
+      printf("TDeclNameRegistry Destructor. List of %lu names:\n",
+             (long unsigned int)fClassNamesSet.size());
+      for (auto const & key: fClassNamesSet) {
+         printf(" - %s\n", key.c_str());
       }
    }
 }
